requirements.txt
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
openai-whisper==20231117
torch==2.1.2
numpy==1.26.2
